#109 feat: 친구 관리 타입 상수 추가
