+++ src/constants/menu/index.ts
@@ -1,4 +1,9 @@
 import { BOTTOM_MENU_INFO } from './bottom'
 import { MENU_INFO as HOME_MENU_INFO } from './home'
-import { FRIENDS_MENU_INFO } from './friends'
-export { BOTTOM_MENU_INFO, HOME_MENU_INFO, FRIENDS_MENU_INFO }
+import { FRIENDS_MENU_INFO, FRIENDS_MENU_INFO_OBJ } from './friends'
+export {
+  BOTTOM_MENU_INFO,
+  HOME_MENU_INFO,
+  FRIENDS_MENU_INFO,
+  FRIENDS_MENU_INFO_OBJ,
+}
