--- src/types/styles/bottomMenu.ts
@@ -1,8 +1,0 @@
-export interface BottomMenuInfoProps {
-  link: string
-  image: {
-    width: number
-    height: number
-    icon: string
-  }
-}
